+++ Sources/SwiftUIKit/Views/List/ListButtonStyle.swift
@@ -9,30 +9,68 @@
 import SwiftUI
 
 
-   /**
-    This button style makes the button label take up the entire
-    width and applies a content shape that makes the entire row
-    tappable.
+/**
+ This button style makes the button label take up the entire
+ width and applies a content shape that makes the entire row
+ tappable.
 
-    You can apply this button style to an entire `List` to make
-    apply this button style to all buttons in the list. You can
-    also use the `buttonStyle(.list)` shorthand.
-    */
-   public struct ListButtonStyle: ButtonStyle {
+ You can apply this button style to an entire `List` to make
+ apply this button style to all buttons in the list. You can
+ also use the `buttonStyle(.list)` shorthand.
+ */
+public struct ListButtonStyle: ButtonStyle {
 
-       public func makeBody(configuration: Configuration) -> some View {
-           configuration.label
-               .frame(maxWidth: .infinity, alignment: .leading)
-               .contentShape(Rectangle())
-       }
-   }
+    public func makeBody(configuration: Configuration) -> some View {
+        configuration.label
+            .frame(maxWidth: .infinity, alignment: .leading)
+            .contentShape(Rectangle())
+    }
+}
 
-   public extension ButtonStyle where Self == ListButtonStyle {
+public extension ButtonStyle where Self == ListButtonStyle {
 
-       /**
-        This style makes the button label take up all available
-        width and also applies a rectangular content shape that
-        makes the entire button row tappable.
-        */
-       static var list: ListButtonStyle { ListButtonStyle() }
-   }
+    /**
+     This style makes the button label take up all available
+     width and also applies a rectangular content shape that
+     makes the entire button row tappable.
+     */
+    static var list: ListButtonStyle { ListButtonStyle() }
+}
+
+struct ListButtonStyle_Previews: PreviewProvider {
+
+    struct Preview: View {
+
+        @State
+        private var overlayText = ""
+
+        var body: some View {
+            List {
+                ForEach(0...100, id: \.self) {
+                    button("Button \($0)")
+                }
+            }
+            .overlay(overlay)
+            .buttonStyle(.list)
+        }
+
+        var overlay: some View {
+            Text(overlayText)
+                .padding()
+                .background(Color.yellow)
+                .cornerRadius(10)
+                .opacity(overlayText.isEmpty ? 0 : 1)
+                .onTapGesture { overlayText = "" }
+        }
+
+        func button(_ name: String) -> some View {
+            Button(name) {
+                overlayText = "\(name) tapped!"
+            }
+        }
+    }
+
+    static var previews: some View {
+        Preview()
+    }
+}